refactor: test route (#653)
* refactor: test

* fix: params type

* refactor: comments

* fix: casing

* chore: comments

* fix: first section function position

* chore: comment resolved

* fix: function

* chore: comment resolved

---------
